Develop

+++ NeoLemmixSharp.Engine/LevelBuilding/Gadgets/HitBoxGadgetArchetypeBuilder.cs
@@ -226,7 +226,7 @@ public sealed class HitBoxGadgetArchetypeBuilder : IGadgetArchetypeBuilder
                 HitBoxType.Rectangular => CreateRectangularHitBoxRegion(gadgetData, item.HitBoxDefinitionData),
                 HitBoxType.PointSet => CreatePointSetHitBoxRegion(gadgetData, item.HitBoxDefinitionData),
 
-                _ => Helpers.ThrowUnknownEnumValueException<HitBoxType, IHitBoxRegion>((int)item.HitBoxType)
+                _ => Helpers.ThrowUnknownEnumValueException<HitBoxType, IHitBoxRegion>(item.HitBoxType)
             };
 
             result.Add(item.Orientation, hitBoxRegion);
